Instant Messaging Policy
Internal policy for employees, contractors, or other third parties to adhere to when participating in company approved chatrooms.  This is just a framework, and should be modified to fit your company guidelines. Please for the love of god, this is not legal advice.  If you have any suggestions for improvements let me know.  I did not want to go into too much detail on security since that should be a separate policy.