Mutual NDA
Warning: All NDAs follow this similar pattern where theres an expectation of privacy regarding what you’re discussing.  As a practical reality, enforcing NDA’s are a nightmare.  Because you end up spending 80% of the time trying to litigate what, if anything, was actually confidential.  Because theres a general exception that something that can be found outside of the nda may not be protectable within the context of the nda communications.  To solve this, you basically have to mark EVERTYHING as “confdiential”.  At the end of the day, depending on the power struggle one of the parties will look at your nda and laugh at it and kick you out of their office.